Fix issue with the decompressor truncating the output in some cases.
If the source reader returns EOF along with the final data block,
instead of returning EOF on the read *after* the final block then that
EOF was getting passed through to the caller immediately. If there was
still decompressed data that hadn't been read yet then this would result
in the decompressed output being truncated.